refactor: split FilteredSetsPanel and dedupe active-scope mutations
Decompose FilteredSetsPanel.tsx (154 lines, four mixed concerns) into
AuthedFilteredSetsPanel/GroupScopedSetsPanel/SetsPanelContent, matching
the earlier ActiveGroupSwitcher/ScopeMenuBody split. This also removes
the duplicated "everyone" SetsPanelContent call between the anonymous
and authed-non-group branches.

Also collapse useSetActiveGroupMutation/useSetActiveScopeMutation into
a shared useProfileFieldMutation, and dedupe the ToggleGroupItem
className repeated in ActiveGroupSetting/ActiveScopeSetting.
