+++ src/pages/EditionView/tabs/VoteTab/FilteredSetsPanel.tsx
@@ -1,18 +1,11 @@
-import { useMemo, useState } from "react";
-import { useSuspenseQuery } from "@tanstack/react-query";
-import { SetsPanel } from "@/pages/EditionView/tabs/VoteTab/SetsPanel";
-import { useSetFiltering } from "@/pages/EditionView/tabs/VoteTab/useSetFiltering";
+import { AuthedFilteredSetsPanel } from "@/pages/EditionView/tabs/VoteTab/AuthedFilteredSetsPanel";
+import { EveryoneSetsPanel } from "@/pages/EditionView/tabs/VoteTab/SetsPanelContent";
 import { FilterSortControls } from "@/pages/EditionView/tabs/VoteTab/filters/FilterSortControls";
-import { groupMembersQuery } from "@/api/groups/useGroupMembers";
 import { useAuth } from "@/contexts/AuthContext";
-import { useActiveScope } from "@/contexts/ActiveScopeContext";
 import type { FestivalSet } from "@/api/sets/types";
 import type { FilterSortState } from "@/hooks/useUrlState";
-import type { BinaryVoteScope, VoteScope } from "@/lib/voteScope";
 
-const NO_MEMBERS = new Set<string>();
-
-interface FilteredSetsPanelProps {
+export interface FilteredSetsPanelProps {
   sets: FestivalSet[];
   urlState: FilterSortState;
   updateUrlState: (updates: Partial<FilterSortState>) => void;
@@ -33,12 +26,10 @@ export function FilteredSetsPanel(props: FilteredSetsPanelProps) {
           editionId={props.editionId}
         />
         <div className="mt-8">
-          <SetsPanelContent
+          <EveryoneSetsPanel
             sets={props.sets}
             urlState={props.urlState}
             updateUrlState={props.updateUrlState}
-            voteScope="everyone"
-            groupMemberIds={NO_MEMBERS}
           />
         </div>
       </>
@@ -46,109 +37,4 @@ export function FilteredSetsPanel(props: FilteredSetsPanelProps) {
   }
 
   return <AuthedFilteredSetsPanel {...props} userId={user.id} />;
-}
-
-function AuthedFilteredSetsPanel(
-  props: FilteredSetsPanelProps & { userId: string },
-) {
-  const { current, groups, activeGroupId } = useActiveScope();
-  const activeGroupName = activeGroupId
-    ? groups.find((group) => group.id === activeGroupId)?.name
-    : undefined;
-  const [perspective, setPerspective] = useState<BinaryVoteScope>(
-    current.kind === "group" ? "group" : "everyone",
-  );
-
-  const voteScope: VoteScope =
-    perspective === "group" && activeGroupId ? "group" : "everyone";
-
-  return (
-    <>
-      <FilterSortControls
-        state={props.urlState}
-        onStateChange={props.updateUrlState}
-        onClear={props.clearFilters}
-        editionId={props.editionId}
-        votePerspective={
-          activeGroupId && activeGroupName
-            ? {
-                scope: voteScope,
-                onScopeChange: setPerspective,
-                groupName: activeGroupName,
-              }
-            : undefined
-        }
-      />
-
-      <div className="mt-8">
-        {voteScope === "group" && activeGroupId ? (
-          <GroupScopedSetsPanel
-            sets={props.sets}
-            urlState={props.urlState}
-            updateUrlState={props.updateUrlState}
-            groupId={activeGroupId}
-          />
-        ) : (
-          <SetsPanelContent
-            sets={props.sets}
-            urlState={props.urlState}
-            updateUrlState={props.updateUrlState}
-            voteScope="everyone"
-            groupMemberIds={NO_MEMBERS}
-          />
-        )}
-      </div>
-    </>
-  );
-}
-
-function GroupScopedSetsPanel({
-  sets,
-  urlState,
-  updateUrlState,
-  groupId,
-}: Pick<FilteredSetsPanelProps, "sets" | "urlState" | "updateUrlState"> & {
-  groupId: string;
-}) {
-  const { data: members } = useSuspenseQuery(groupMembersQuery(groupId));
-  const groupMemberIds = useMemo(
-    () => new Set(members.map((member) => member.user_id)),
-    [members],
-  );
-
-  return (
-    <SetsPanelContent
-      sets={sets}
-      urlState={urlState}
-      updateUrlState={updateUrlState}
-      voteScope="group"
-      groupMemberIds={groupMemberIds}
-    />
-  );
-}
-
-function SetsPanelContent({
-  sets,
-  urlState,
-  updateUrlState,
-  voteScope,
-  groupMemberIds,
-}: Pick<FilteredSetsPanelProps, "sets" | "urlState" | "updateUrlState"> & {
-  voteScope: VoteScope;
-  groupMemberIds: Set<string>;
-}) {
-  const { filteredAndSortedSets, lockCurrentOrder } = useSetFiltering(
-    sets,
-    urlState,
-    voteScope,
-    groupMemberIds,
-  );
-
-  return (
-    <SetsPanel
-      sets={filteredAndSortedSets}
-      use24Hour={urlState.use24Hour}
-      onLockSort={() => lockCurrentOrder(updateUrlState)}
-    />
-  );
 }
