+++ src/pages/Settings/ActiveGroupSetting.tsx
@@ -1,6 +1,7 @@
 import { Users } from "lucide-react";
 import { ToggleGroup, ToggleGroupItem } from "@/components/ui/toggle-group";
 import { useActiveScope } from "@/contexts/ActiveScopeContext";
+import { settingsToggleItemClassName } from "@/pages/Settings/settingsToggleItemClassName";
 
 export function ActiveGroupSetting() {
   const { groups, activeGroupId, setActiveGroup } = useActiveScope();
@@ -26,7 +27,7 @@ export function ActiveGroupSetting() {
           <ToggleGroupItem
             key={group.id}
             value={group.id}
-            className="gap-1.5 rounded-md border border-purple-400/30 px-3 py-1.5 text-sm text-purple-100 data-[state=on]:border-purple-400 data-[state=on]:bg-purple-600/20 data-[state=on]:font-medium data-[state=on]:text-purple-100 hover:bg-purple-600/10"
+            className={settingsToggleItemClassName}
             aria-label={`Set ${group.name} as active group`}
           >
             <Users className="h-3.5 w-3.5" />
