+++ src/pages/Settings/settingsToggleItemClassName.ts
@@ -1,0 +1,2 @@
+export const settingsToggleItemClassName =
+  "gap-1.5 rounded-md border border-purple-400/30 px-3 py-1.5 text-sm text-purple-100 data-[state=on]:border-purple-400 data-[state=on]:bg-purple-600/20 data-[state=on]:font-medium data-[state=on]:text-purple-100 hover:bg-purple-600/10";
